+++ rdata/unparser/_xdr.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from rdata.parser import R_INT_NA
 
-from .base import Writer
+from ._unparser import Writer
 
 if TYPE_CHECKING:
     import io
